+++ scripts/bladebit_setup.sh
@@ -23,11 +23,9 @@ if [[ (${mode} =~ ^fullnode.*  || ${mode} =~ "plotter") && (${blockchains} == 'c
             cmake --build . --target bladebit --config Release
             ln -s /bladebit/build/bladebit /usr/bin/bladebit
             cd / && echo "Bladebit version: "`bladebit --version`
-            # Now go get the Bladebit closed source binary of 3.0.0 alpha, very securely zipped twice :)
             cd /bladebit/build
-            curl -sLJO https://download.chia.net/bladebit/alpha3/bladebit-cuda-v3.0.0-alpha3-ubuntu-x86-64.tar.gz.zip
-            unzip bladebit-cuda-v3.0.0-alpha3-ubuntu-x86-64.tar.gz.zip
-            tar -xzf bladebit-cuda-v3.0.0-alpha3-ubuntu-x86-64.tar.gz
+            curl -sLJO https://download.chia.net/bladebit/alpha4/bladebit-cuda-plotter/bladebit-cuda-v3.0.0-alpha4-ubuntu-x86-64.tar.gz
+            tar -xzf bladebit-cuda-v3.0.0-alpha4-ubuntu-x86-64.tar.gz
             chmod 755 bladebit_cuda
             ln -s /bladebit/build/bladebit_cuda /usr/bin/bladebit_cuda
         else